refactor(session-monitor): extract registry queries into session-monitor.queries.lib.mjs
session-monitor.lib.mjs was 1416 lines, past the 1350-line early-warning
threshold the CI file-headroom check enforces (issue #1593).

The six read-only registry queries (hasActiveSessionForUrl,
findStoppableSessionByUrl, hasActiveSessionForUrlAsync,
getRunningTrackedIsolationSessions, getRunningSessionItems, getSessionStats)
only need the registry Map plus the liveness helpers, so they move behind a
factory that receives those as injected dependencies. They are re-exported
unchanged, so every importer and test keeps working.

session-monitor.lib.mjs: 1416 -> 1156 lines.

Refs #2175
